requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.6
django-rosetta==0.9.4
idna==2.9
Pillow==7.1.2
polib==1.1.0
protobuf==3.12.1
psycopg2-binary==2.8.5
python-gettext==4.0
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
